Move Fungible ABI to linera SDK
