+++ linera-sdk/src/abis/mod.rs
@@ -1,0 +1,6 @@
+// Copyright (c) Zefchain Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+//! Common ABIs that may have multiple implementations.
+
+pub mod fungible;
