+++ linera-service/Cargo.toml
@@ -13,7 +13,7 @@ edition = "2021"
 [features]
 default = ["wasmer", "rocksdb"]
 test = ["linera-views/test", "linera-execution/test"]
-benchmark = ["linera-base/test", "dep:fungible", "dep:port-selector"]
+benchmark = ["linera-base/test", "dep:linera-sdk", "dep:port-selector"]
 wasmer = ["linera-execution/wasmer", "linera-storage/wasmer"]
 wasmtime = ["linera-execution/wasmtime", "linera-storage/wasmtime"]
 rocksdb = [
@@ -50,7 +50,6 @@ dirs.workspace = true
 file-lock = "2.1.11"
 fs-err.workspace = true
 fs_extra = { workspace = true, optional = true }
-fungible = { workspace = true, optional = true }
 futures.workspace = true
 hex.workspace = true
 http.workspace = true
@@ -61,6 +60,7 @@ linera-chain = { workspace = true, features = ["metrics"] }
 linera-core = { workspace = true, features = ["metrics", "rocksdb", "wasmer"] }
 linera-execution = { workspace = true, features = ["fs", "metrics", "wasmer"] }
 linera-rpc = { workspace = true, features = ["server", "simple-network"] }
+linera-sdk = { workspace = true, optional = true }
 linera-storage = { workspace = true, features = ["metrics", "wasmer"] }
 linera-storage-service.workspace = true
 linera-version.workspace = true
